p2s_requirements.txt
absl-py==1.4.0
array-record==0.5.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
compress-json==1.0.10
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
einops==0.7.0
etils==1.6.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
joblib==1.3.2
json5==0.9.14
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-collections==0.1.1
ml-dtypes==0.2.0
netifaces==0.11.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
paramparse==1.8.6
Pillow==10.1.0
promise==2.3
protobuf==3.20.3
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scikit-video==1.1.11
scipy==1.11.4
six==1.16.0
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-gan==2.1.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.23.0
tensorflow-text==2.15.0
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.12.9
toml==0.10.2
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
